+++ alf/algorithms/agent.py
@@ -114,6 +114,8 @@ class Agent(OnPolicyAlgorithm):
             reward_shaping_fn=reward_shaping_fn,
             observation_transformer=observation_transformer,
             debug_summaries=debug_summaries,
+            summarize_grads_and_vars=summarize_grads_and_vars,
+            summarize_action_distributions=summarize_action_distributions,
             name=name)
 
         self._rl_algorithm = rl_algorithm
