+++ alf/algorithms/sac_algorithm.py
@@ -99,6 +99,8 @@ class SacAlgorithm(OffPolicyAlgorithm):
                  alpha_optimizer=None,
                  gradient_clipping=None,
                  debug_summaries=False,
+                 summarize_grads_and_vars=False,
+                 summarize_action_distributions=False,
                  name="SacAlgorithm"):
         """Create a SacAlgorithm
 
@@ -124,6 +126,9 @@ class SacAlgorithm(OffPolicyAlgorithm):
             alpha_optimizer (tf.optimizers.Optimizer): The optimizer for alpha.
             gradient_clipping (float): Norm length to clip gradients.
             debug_summaries (bool): True if debug summaries should be created.
+            summarize_grads_and_vars (bool): If True, gradient and network
+                variable summaries will be written during training.
+            summarize_action_distributions (bool): If True, generate summaris
             name (str): The name of this algorithm.
         """
         critic_network1 = critic_network
@@ -152,6 +157,8 @@ class SacAlgorithm(OffPolicyAlgorithm):
                                    [log_alpha]],
             gradient_clipping=gradient_clipping,
             debug_summaries=debug_summaries,
+            summarize_grads_and_vars=summarize_grads_and_vars,
+            summarize_action_distributions=summarize_action_distributions,
             name=name)
 
         self._log_alpha = log_alpha
