+++ alf/drivers/policy_driver.py
@@ -16,7 +16,7 @@ import abc
 import os
 from typing import Callable
 
-import psutil
+# import psutil
 import math
 import gin.tf
 import tensorflow as tf
@@ -40,15 +40,12 @@ class PolicyDriver(driver.Driver):
                  metrics=[],
                  training=True,
                  greedy_predict=False,
-                 debug_summaries=False,
-                 summarize_grads_and_vars=False,
-                 summarize_action_distributions=False,
                  train_step_counter=None):
         """Create a PolicyDriver.
 
         Args:
             env (TFEnvironment): A TFEnvoronmnet
-            algorithm (OnPolicyAlgorith): The algorithm for training
+            algorithm (RLAlgorithm): The algorithm for training
             observers (list[Callable]): An optional list of observers that are
                 updated after every step in the environment. Each observer is a
                 callable(time_step.Trajectory).
@@ -58,11 +55,6 @@ class PolicyDriver(driver.Driver):
             training (bool): True for training, false for evaluating
             greedy_predict (bool): use greedy action for evaluation (i.e.
                 training==False).
-            debug_summaries (bool): A bool to gather debug summaries.
-            summarize_grads_and_vars (bool): If True, gradient and network
-                variable summaries will be written during training.
-            summarize_action_distributions (bool): If True, generate summaris
-                for the action distributions.
             train_step_counter (tf.Variable): An optional counter to increment
                 every time the a new iteration is started. If None, it will use
                 tf.summary.experimental.get_step(). If this is still None, a
@@ -88,7 +80,6 @@ class PolicyDriver(driver.Driver):
                     batch_size=env.batch_size, buffer_size=metric_buf_size),
             ]
         self._metrics = standard_metrics + metrics
-        self._exp_observers = []
         self._use_rollout_state = use_rollout_state
 
         super(PolicyDriver, self).__init__(env, None,
@@ -97,59 +88,13 @@ class PolicyDriver(driver.Driver):
         self._algorithm = algorithm
         self._training = training
         self._greedy_predict = greedy_predict
-        self._debug_summaries = debug_summaries
-        self._summarize_grads_and_vars = summarize_grads_and_vars
-        self._summarize_action_distributions = summarize_action_distributions
         self._train_step_counter = common.get_global_counter(
             train_step_counter)
-        self._proc = psutil.Process(os.getpid())
         if training:
             self._policy_state_spec = algorithm.train_state_spec
         else:
             self._policy_state_spec = algorithm.predict_state_spec
         self._initial_state = self.get_initial_policy_state()
-
-    def add_experience_observer(self, observer: Callable):
-        """Add an observer to receive experience.
-
-        Args:
-            observer (Callable): callable which accept Experience as argument.
-        """
-        self._exp_observers.append(observer)
-
-    def _training_summary(self, training_info, loss_info, grads_and_vars):
-        if self._summarize_grads_and_vars:
-            summary_utils.add_variables_summaries(grads_and_vars,
-                                                  self._train_step_counter)
-            summary_utils.add_gradients_summaries(grads_and_vars,
-                                                  self._train_step_counter)
-        if self._debug_summaries:
-            common.add_action_summaries(training_info.action,
-                                        self.env.action_spec())
-            common.add_loss_summaries(loss_info)
-
-        if self._summarize_action_distributions:
-            summary_utils.summarize_action_dist(
-                training_info.action_distribution, self.env.action_spec())
-            if training_info.collect_action_distribution:
-                summary_utils.summarize_action_dist(
-                    action_distributions=training_info.
-                    collect_action_distribution,
-                    action_specs=self.env.action_spec(),
-                    name="collect_action_dist")
-
-        for metric in self.get_metrics():
-            metric.tf_summaries(
-                train_step=self._train_step_counter,
-                step_metrics=self.get_metrics()[:2])
-
-        mem = tf.py_function(
-            lambda: self._proc.memory_info().rss // 1e6, [],
-            tf.float32,
-            name='memory_usage')
-        if not tf.executing_eagerly():
-            mem.set_shape(())
-        tf.summary.scalar(name='memory_usage', data=mem)
 
     @property
     def algorithm(self):
@@ -223,7 +168,7 @@ class PolicyDriver(driver.Driver):
                                    next_time_step)
             for observer in self._observers:
                 observer(traj)
-        if self._exp_observers:
+        if self._algorithm.exp_observers and self._training:
             action_distribution_param = common.get_distribution_params(
                 policy_step.action)
             exp = make_experience(
@@ -231,7 +176,7 @@ class PolicyDriver(driver.Driver):
                 policy_step._replace(action=action),
                 action_distribution=action_distribution_param,
                 state=policy_state if self._use_rollout_state else ())
-            for observer in self._exp_observers:
+            for observer in self._algorithm.exp_observers:
                 observer(exp)
 
         return next_time_step, policy_step, action, transformed_time_step
