+++ alf/drivers/sync_off_policy_driver.py
@@ -55,8 +55,6 @@ class SyncOffPolicyDriver(OffPolicyDriver):
                  observers=[],
                  use_rollout_state=False,
                  metrics=[],
-                 debug_summaries=False,
-                 summarize_grads_and_vars=False,
                  train_step_counter=None):
         """Create an OffPolicyDriver.
 
@@ -71,9 +69,6 @@ class SyncOffPolicyDriver(OffPolicyDriver):
             use_rollout_state (bool): Include the RNN state for the experiences
                 used for off-policy training
             metrics (list[TFStepMetric]): An optiotional list of metrics.
-            debug_summaries (bool): A bool to gather debug summaries.
-            summarize_grads_and_vars (bool): If True, gradient and network
-                variable summaries will be written during training.
             train_step_counter (tf.Variable): An optional counter to increment
                 every time the a new iteration is started. If None, it will use
                 tf.summary.experimental.get_step(). If this is still None, a
@@ -90,8 +85,6 @@ class SyncOffPolicyDriver(OffPolicyDriver):
             observers=observers,
             use_rollout_state=use_rollout_state,
             metrics=metrics,
-            debug_summaries=debug_summaries,
-            summarize_grads_and_vars=summarize_grads_and_vars,
             train_step_counter=train_step_counter)
 
     def _run(self, max_num_steps, time_step, policy_state):
